update transmission oplife
