+++ workflow/scripts/osemosys_global/visualise.py
@@ -383,3 +383,54 @@ def plot_transmission_flow(
 
 if __name__ == '__main__':
     main()
+
+
+'''
+# ### Interactive visualisation of residual capacity by node
+
+import matplotlib.pyplot as plt
+import seaborn as sns; sns.set(color_codes = True)
+from ipywidgets import interact, interactive, fixed, interact_manual, Layout
+import ipywidgets as widgets
+#importing plotly and cufflinks in offline mode
+import plotly as py
+#import plotly.graph_objs as go
+import cufflinks
+import plotly.offline as pyo
+from plotly.offline import plot, iplot, init_notebook_mode
+pyo.init_notebook_mode()
+cufflinks.go_offline()
+cufflinks.set_config_file(world_readable=True, theme='white')
+
+color_codes = pd.read_csv(r'data\color_codes.csv', encoding='latin-1')
+color_dict = dict([(n,c) for n,c in zip(color_codes.tech, color_codes.colour)])
+
+def f(node):
+    df_plot = df_res_cap_plot.loc[df_res_cap_plot['node_code']==node]
+    df_plot.drop('node_code', 
+                        axis = 1, 
+                        inplace = True)
+    df_plot = df_plot.pivot_table(index='model_year',
+                                    columns='tech_code',
+                                    values='value',
+                                    aggfunc='sum').reset_index()
+
+
+    #plt.figure(figsize=(10, 10), dpi= 80, facecolor='w', edgecolor='k')
+    #ax = sns.barplot(df_plot)
+    return df_plot.iplot(x = 'model_year',
+                            kind = 'bar', 
+                            barmode = 'stack',
+                            xTitle = 'Year',
+                            yTitle = 'Gigawatts',
+                            color=[color_dict[x] for x in df_plot.columns if x != 'model_year'],
+                            title = 'Residual Capacity',
+                            showlegend = True)
+
+interact(f,
+            node=widgets.Dropdown(options = (df_res_cap_plot['node_code']
+                                            .unique()
+                                            )
+                                )
+        )
+'''
